requirements.txt
fastapi==0.115.12
numpy<2.0.0
torch==2.2.0
torchvision==0.17.0
pillow==10.2.0
python-multipart==0.0.9
uvicorn==0.27.1
